Make test functions reused instead of duplicated and fix comment
