+++ src/query/models/test/tags.go
@@ -1,0 +1,41 @@
+// Copyright (c) 2019 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package test
+
+import "github.com/m3db/m3/src/query/models"
+
+// MustMakeTags creates tags given that the number of args is even.
+func MustMakeTags(tag ...string) models.Tags {
+	if len(tag)%2 != 0 {
+		panic("must have even tag length")
+	}
+
+	tagLength := len(tag) / 2
+	t := models.NewTags(tagLength, models.NewTagOptions())
+	for i := 0; i < tagLength; i++ {
+		t = t.AddTag(models.Tag{
+			Name:  []byte(tag[i*2]),
+			Value: []byte(tag[i*2+1]),
+		})
+	}
+
+	return t
+}
