+++ src/query/test/block.go
@@ -234,21 +234,3 @@ func GenerateValuesAndBounds(
 
 	return values, bounds
 }
-
-// MustMakeTags creates tags given that the number of args is even.
-func MustMakeTags(tag ...string) models.Tags {
-	if len(tag)%2 != 0 {
-		panic("must have even tag length")
-	}
-
-	tagLength := len(tag) / 2
-	t := models.NewTags(tagLength, models.NewTagOptions())
-	for i := 0; i < tagLength; i++ {
-		t = t.AddTag(models.Tag{
-			Name:  []byte(tag[i*2]),
-			Value: []byte(tag[i*2+1]),
-		})
-	}
-
-	return t
-}
